Some file restructuring
