+++ include/Base.hpp
@@ -1,7 +1,6 @@
 #ifndef EVK_hpp
 #define EVK_hpp
 
-//#define MSAA
 
 #include <iostream>
 #include <assert.h>
@@ -11,147 +10,19 @@
 #include <set>
 #include <optional>
 
-#include <MoltenVK/mvk_vulkan.h>
-#include <vma/vk_mem_alloc.h>
-#include <mattresses.h>
 
-#include <SDL2/SDL_image.h>
 
-#include <Static.hpp>
-
-#define MAX_FRAMES_IN_FLIGHT 2
-
-#define VULKAN_VERSION_USING VK_API_VERSION_1_2
-
-/*
- !!! BEWARE !!!
-	 Be careful about the scopes of large blueprints that contain Vulkan 'create info's of various kinds.
-	 If a Vulkan 'create info' has a pointer to some stuff, that pointer must remain valid at the point
-	 of use of the outer-most blueprint!
- */
-
-/*
- Initialisation:
-	\/ For with SDL
-	- `SDL_Init(...)`
-	- `SDL_CreateWindow(..., SDL_WINDOW_VULKAN | SDL_WINDOW_RESIZABLE | ...)`
-	/\
-	- Create an instance of `EVK::Devices`
-		\/ For with SDL
-		- Surface creation function:
-`[window](VkInstance instance) -> VkSurfaceKHR {
-	VkSurfaceKHR ret;
-	if(SDL_Vulkan_CreateSurface(window, instance, &ret) == SDL_FALSE)
-		throw std::runtime_error("failed to create window devices.surface!");
-	return ret;
-}`
-		- Extent getter function:
- `[window]() -> VkExtent2D {
-	 int width, height;
-	 SDL_GL_GetDrawableSize(window, &width, &height);
-	 return (VkExtent2D){
-		 .width = uint32_t(width),
-		 .height = uint32_t(height)
-	 };
- }`
-		/\
-	- Create an instance of `EVK::Interface`
-	- Add a window resize event callback to call method `EVK::Interface::FramebufferResizeCallback()` of the interface
-		\/ For with SDL
-		-   SDL_Event event;
-			event.type = SDL_WINDOWEVENT;
-			event.window.event = SDL_WINDOWEVENT_SIZE_CHANGED;
-		/\
-		\/ For with wxWidgets
-		- Event type: `wxEVT_SIZE`
-		/\
-	- Fill required vertex and index buffers with `Vulkan::FillVertexBuffer` / `Vulkan::FillIndexBuffer`
-	- Build structuers like UBOs, textures and buffered render passes
-	- Build pipelines
- 
- In the render loop:
-	- (Access render pipelines with `EVK::Interface::RP(...)`, and pipeline descriptor sets with `EVK::Interface::RenderPipeline::DS(...)`)
- 
-	- Set the values of your uniforms by modifying the values pointed to by that returned/filled by `EVK::Interface::GetUniformBufferObjectPointer<<UBO struct type>>(...)` / `EVK::Interface::GetUniformBufferObjectPointers<<UBO struct type>>(...)`
-	- Put your render pass in the following condition: `if(<EVK::Interface instance>.BeginFrame()){ ... }`
-	- For each render pass:
-		- Begin it with `EVK::Interface::CmdBegin(Layered)BufferedRenderPass(...)`, or `EVK::Interface::BeginFinalRenderPass()` for the final one
-		- For each render pipeline:
-			- Bind with `EVK::Interface::RenderPipeline::Bind()`
-			Then, as many times as you want:
-				- Set the desired descriptor set bindings with `EVK::Interface::RenderPipeline::BindDescriptorSets(...)` (this includes setting dynamic UBO offsets)
-				- Set the desired push constant values with `EVK::Interface::RenderPipeline::CmdPushConstants<<PC struct type>>(...)`
-				- Queue a draw call command `EVK::Interface::CmdDraw(...)` or `EVK::Interface::CmdDrawIndexed(...)`
-		- Finish the pass with `EVK::Interface::CmdEndRenderPass()`, or `EVK::Interface::EndFinalRenderPassAndFrame()` for the final one
- 
- Cleaning up:
-	- `vkDeviceWaitIdle(EVK::Interface::GetLogicalDevice())`
-	- Destroy your `EVK::Interface` instance
-	- `SDL_DestroyWindow(<your SDL_Window pointer>)`
- */
 
 namespace EVK {
 
 class Interface;
 class Devices;
 
-struct SwapChainSupportDetails {
-	VkSurfaceCapabilitiesKHR capabilities;
-	std::vector<VkSurfaceFormatKHR> formats;
-	std::vector<VkPresentModeKHR> presentModes;
-};
-struct QueueFamilyIndices {
-	std::optional<uint32_t> graphicsAndComputeFamily;
-	std::optional<uint32_t> presentFamily;
-	
-	bool IsComplete(){
-		return graphicsAndComputeFamily.has_value() && presentFamily.has_value();
-	}
-};
+
 
 //VkFormat SDLPixelFormatToVulkanFormat(const SDL_PixelFormatEnum &sdlPixelFormat);
 
-struct VertexBufferObject {
-	VkBuffer bufferHandle;
-	VkDeviceSize offset;
-	VmaAllocation allocation;
-	VkDeviceSize size;
-	
-	void CleanUp(const VmaAllocator &allocator){
-		vmaDestroyBuffer(allocator, bufferHandle, allocation);
-	}
-};
-struct IndexBufferObject {
-	VkBuffer bufferHandle;
-	VkDeviceSize offset;
-	VmaAllocation allocation;
-	uint32_t indexCount;
-	
-	void CleanUp(const VmaAllocator &allocator){
-		vmaDestroyBuffer(allocator, bufferHandle, allocation);
-	}
-};
 
-struct UniformBufferObject {
-	VkBuffer buffersFlying[MAX_FRAMES_IN_FLIGHT];
-	VmaAllocation allocationsFlying[MAX_FRAMES_IN_FLIGHT];
-	VmaAllocationInfo allocationInfosFlying[MAX_FRAMES_IN_FLIGHT];
-	VkDeviceSize size;
-	
-	void CleanUp(const VmaAllocator &allocator){
-		for(int i=0; i<MAX_FRAMES_IN_FLIGHT; i++) vmaDestroyBuffer(allocator, buffersFlying[i], allocationsFlying[i]);
-	}
-};
-struct StorageBufferObject {
-	VkBuffer buffersFlying[MAX_FRAMES_IN_FLIGHT];
-	VmaAllocation allocationsFlying[MAX_FRAMES_IN_FLIGHT];
-	VmaAllocationInfo allocationInfosFlying[MAX_FRAMES_IN_FLIGHT];
-	VkDeviceSize size;
-	
-	void CleanUp(const VmaAllocator &allocator){
-		for(int i=0; i<MAX_FRAMES_IN_FLIGHT; i++) vmaDestroyBuffer(allocator, buffersFlying[i], allocationsFlying[i]);
-	}
-};
 
 struct PNGImageBlueprint {
 	std::string imageFilename;
@@ -182,123 +53,8 @@ struct ManualImageBlueprint {
 	VkMemoryPropertyFlags properties;
 	VkImageAspectFlags aspectFlags;
 };
-struct TextureImage {
-	VkImage image;
-	VmaAllocation allocation;
-	VkImageView view;
-	uint32_t mipLevels;
-	ManualImageBlueprint blueprint;
-	
-	void CleanUp(const VkDevice &logicalDevice, const VmaAllocator &allocator){
-		vkDestroyImageView(logicalDevice, view, nullptr);
-		vmaDestroyImage(allocator, image, allocation);
-	}
-};
-struct BufferedRenderPass {
-	VkRenderPass renderPass;
-	std::vector<int> targetTextureImageIndices;
-	VkFramebuffer frameBuffersFlying[MAX_FRAMES_IN_FLIGHT];
-	std::optional<vec<2, uint32_t>> size;
-	
-	void CleanUp(const VkDevice &logicalDevice){
-		vkDestroyRenderPass(logicalDevice, renderPass, nullptr);
-		for(int i=0; i<MAX_FRAMES_IN_FLIGHT; i++) vkDestroyFramebuffer(logicalDevice, frameBuffersFlying[i], nullptr);
-	}
-};
-struct LayeredBufferedRenderPass {
-	LayeredBufferedRenderPass(int layersN) {
-		layers = std::vector<Layer>(layersN);
-	}
-	
-	uint32_t width, height;
-	VkRenderPass renderPass;
-	int targetTextureImageIndex;
-	VkFormat imageFormat;
-	VkImageAspectFlags imageAspectFlags;
-	
-	struct Layer {
-		VkImageView imageView;
-		VkFramebuffer frameBuffersFlying[MAX_FRAMES_IN_FLIGHT];
-		
-		void CleanUp(const VkDevice &logicalDevice){
-			vkDestroyImageView(logicalDevice, imageView, nullptr);
-			for(int i=0; i<MAX_FRAMES_IN_FLIGHT; i++){
-				vkDestroyFramebuffer(logicalDevice, frameBuffersFlying[i], nullptr);
-			}
-		}
-	};
-	std::vector<Layer> layers;
-	
-	void CleanUp(const VkDevice &logicalDevice){
-		vkDestroyRenderPass(logicalDevice, renderPass, nullptr);
-		for(Layer &layer : layers) layer.CleanUp(logicalDevice);
-	}
-};
 
-class Devices {
-public:
-	Devices(const char *applicationName, std::vector<const char *> requiredExtensions, std::function<VkSurfaceKHR (VkInstance)> surfaceCreationFunction, std::function<VkExtent2D ()> _getExtentFunction);
-	
-	Devices() = delete;
-	Devices(const Devices &) = delete;
-	Devices &operator=(const Devices &) = delete;
-	Devices(Devices &&) = default;
-	Devices &operator=(Devices &&) = default;
-	~Devices();
-	
-	
-	// shader modules
-	VkShaderModule CreateShaderModule(const char *filename) const;
-	
-	VkFormatProperties GetFormatProperties(const VkFormat &format) const {
-		VkFormatProperties ret;
-		vkGetPhysicalDeviceFormatProperties(physicalDevice, format, &ret);
-		return ret;
-	}
-	
-	const VkPhysicalDeviceProperties &GetPhysicalDeviceProperties() const { return physicalDeviceProperties; }
-	
-#ifdef MSAA
-	const VkSampleCountFlagBits &GetMSAASamples() const { return msaaSamples; }
-#endif
-	
-	uint32_t FindMemoryType(uint32_t typeFilter, VkMemoryPropertyFlags properties) const;
-	VkFormat FindSupportedFormat(const std::vector<VkFormat> &candidates, const VkImageTiling &tiling, const VkFormatFeatureFlags &features) const;
-	VkFormat FindDepthFormat() const;
-	
-	SwapChainSupportDetails QuerySwapChainSupport() const;
-	
-	VkExtent2D GetSurfaceExtent() const { return getExtentFunction(); }
-	
-private:
-	void CreateInstance(const char *applicationName, std::vector<const char *> requiredExtensions);
-	void Init();
-	
-	VkInstance instance;
-	VkSurfaceKHR surface;
-	VkPhysicalDevice physicalDevice;
-	VkPhysicalDeviceProperties physicalDeviceProperties;
-	QueueFamilyIndices queueFamilyIndices;
-	VkDevice logicalDevice;
-	VmaAllocator allocator;
-	
-	VkDebugUtilsMessengerEXT debugMessenger;
-	
-	std::function<VkExtent2D ()> getExtentFunction;
-	
-	// queues:
-	VkQueue graphicsQueue;
-	VkQueue presentQueue;
-	VkQueue computeQueue;
-	
-#ifdef MSAA
-	// multi-sampled anti-aliasing (MSAA):
-	VkSampleCountFlagBits msaaSamples = VK_SAMPLE_COUNT_1_BIT; // default to 1 sample per pixel (no multi-sampling)
-	VkSampleCountFlagBits GetMaxUsableSampleCount() const;
-#endif
-	
-	friend class Interface;
-};
+
 
 enum class DescriptorType {
 	UBO,
@@ -326,6 +82,7 @@ struct DescriptorBlueprint {
 	 - combinedImageSampler: {textureSampler indices...}
 	 */
 };
+
 
 using DescriptorSetBlueprint = std::vector<DescriptorBlueprint>;
 
@@ -393,398 +150,6 @@ struct InterfaceBlueprint {
 	int indexBuffersN;
 };
 
-
-
-
-
-// ----------------------------
-// ----- New static stuff -----
-// ----------------------------
-
-// Push constants
-// -----
-struct NoPushConstants {};
-template <VkShaderStageFlags stageFlags, size_t offset, typename T> struct PushConstants {
-	static constexpr VkShaderStageFlags stageFlagsValue = stageFlags;
-	static constexpr uint32_t offsetValue = offset;
-	using type = T;
-};
-
-
-// Descriptor
-// -----
-template <uint32_t binding, VkShaderStageFlags stageFlags> struct DescriptorBase {
-	static constexpr uint32_t bindingValue = binding;
-	static constexpr VkShaderStageFlags stageFlagsValue = stageFlags;
-	
-	virtual VkDescriptorSetLayoutBinding LayoutBinding() const = 0;
-	
-	virtual VkWriteDescriptorSet DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const = 0;
-	
-	virtual VkDescriptorPoolSize PoolSize() const = 0;
-};
-
-template <uint32_t binding, VkShaderStageFlags stageFlags>
-class UBODescriptor : public DescriptorBase<binding, stageFlags> {
-public:
-	UBODescriptor(int _index) : index(_index) {}
-	
-	VkDescriptorSetLayoutBinding LayoutBinding() const override;
-	
-	VkWriteDescriptorSet DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const override;
-	
-	VkDescriptorPoolSize PoolSize() const override;
-	
-private:
-	int index;
-};
-
-template <uint32_t binding, VkShaderStageFlags stageFlags>
-class SBODescriptor : public DescriptorBase<binding, stageFlags> {
-public:
-	SBODescriptor(int _index, int _flightOffset) : index(_index), flightOffset(_flightOffset) {}
-	
-	VkDescriptorSetLayoutBinding LayoutBinding() const override;
-	
-	VkWriteDescriptorSet DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const override;
-	
-	VkDescriptorPoolSize PoolSize() const override;
-	
-private:
-	int index;
-	int flightOffset;
-};
-
-template <uint32_t binding, VkShaderStageFlags stageFlags>
-class TextureImagesDescriptor : public DescriptorBase<binding, stageFlags> {
-public:
-	TextureImagesDescriptor(const std::vector<int> &_indices) : indices(_indices) {}
-	
-	VkDescriptorSetLayoutBinding LayoutBinding() const override;
-	
-	VkWriteDescriptorSet DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const override;
-	
-	VkDescriptorPoolSize PoolSize() const override;
-	
-private:
-	std::vector<int> indices;
-};
-
-template <uint32_t binding, VkShaderStageFlags stageFlags>
-class TextureSamplersDescriptor : public DescriptorBase<binding, stageFlags> {
-public:
-	TextureSamplersDescriptor(const std::vector<int> &_indices) : indices(_indices) {}
-	
-	VkDescriptorSetLayoutBinding LayoutBinding() const override;
-	
-	VkWriteDescriptorSet DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const override;
-	
-	VkDescriptorPoolSize PoolSize() const override;
-	
-private:
-	std::vector<int> indices;
-};
-
-template <uint32_t binding, VkShaderStageFlags stageFlags>
-class CombinedImageSamplersDescriptor : public DescriptorBase<binding, stageFlags> {
-public:
-	CombinedImageSamplersDescriptor(const std::vector<int> &_textureImageIndices, const std::vector<int> &_samplerIndices) : textureImageIndices(_textureImageIndices), samplerIndices(_samplerIndices) {
-		assert(_textureImageIndices.size() == _samplerIndices.size());
-	}
-	
-	VkDescriptorSetLayoutBinding LayoutBinding() const override;
-	
-	VkWriteDescriptorSet DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const override;
-	
-	VkDescriptorPoolSize PoolSize() const override;
-	
-private:
-	std::vector<int> textureImageIndices;
-	std::vector<int> samplerIndices;
-};
-
-template <uint32_t binding, VkShaderStageFlags stageFlags>
-class StorageImagesDescriptor : public DescriptorBase<binding, stageFlags> {
-public:
-	StorageImagesDescriptor(const std::vector<int> &_indices) : indices(_indices) {}
-	
-	VkDescriptorSetLayoutBinding LayoutBinding() const override;
-	
-	VkWriteDescriptorSet DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const override;
-	
-	VkDescriptorPoolSize PoolSize() const override;
-	
-private:
-	std::vector<int> indices;
-};
-
-
-// Uniform
-// -----
-template <uint32_t set, uint32_t binding, VkShaderStageFlags stageFlags> struct UniformBase {
-	static constexpr uint32_t setValue = set;
-	static constexpr uint32_t bindingValue = binding;
-	
-	using descriptorType = DescriptorBase<binding, stageFlags>;
-};
-
-template <uint32_t set, uint32_t binding, VkShaderStageFlags stageFlags, typename T>
-struct UBO : public UniformBase<set, binding, stageFlags> {
-	using descriptorType = UBODescriptor<binding, stageFlags>;
-};
-template <uint32_t set, uint32_t binding, VkShaderStageFlags stageFlags, typename T>
-struct SBO : public UniformBase<set, binding, stageFlags> {
-	using descriptorType = SBODescriptor<binding, stageFlags>;
-};
-template <uint32_t set, uint32_t binding, VkShaderStageFlags stageFlags>
-struct TextureImages : public UniformBase<set, binding, stageFlags> {
-	using descriptorType = TextureImagesDescriptor<binding, stageFlags>;
-};
-template <uint32_t set, uint32_t binding, VkShaderStageFlags stageFlags>
-struct TextureSamplers : public UniformBase<set, binding, stageFlags> {
-	using descriptorType = TextureSamplersDescriptor<binding, stageFlags>;
-};
-template <uint32_t set, uint32_t binding, VkShaderStageFlags stageFlags>
-struct CombinedImageSamplers : public UniformBase<set, binding, stageFlags> {
-	using descriptorType = CombinedImageSamplersDescriptor<binding, stageFlags>;
-};
-template <uint32_t set, uint32_t binding, VkShaderStageFlags stageFlags>
-struct StorageImages : public UniformBase<set, binding, stageFlags> {
-	using descriptorType = StorageImagesDescriptor<binding, stageFlags>;
-};
-
-template <typename uniformT, typename... uniformTs> consteval bool UniformContains(){
-	return ((uniformT::UniformBase::setValue == uniformTs::UniformBase::setValue && uniformT::UniformBase::bindingValue == uniformTs::UniformBase::bindingValue) || ...);
-}
-template <typename uniformT> consteval bool UniformUnique(){ return true; }
-template <typename uniformT1, typename uniformT2, typename... uniformTs> consteval bool UniformUnique(){
-	return !UniformContains<uniformT1, uniformT2, uniformTs...>() && UniformUnique<uniformT2, uniformTs...>();
-}
-
-template <unsigned set, typename uniformT>
-using FilteredDescriptorPackOne = std::conditional_t<uniformT::setValue == set, TypePack<typename uniformT::descriptorType>, TypePack<>>;
-
-template <unsigned set, typename... uniformTs>
-using FilteredDescriptorPack = ConcatenatedPack_t<FilteredDescriptorPackOne<set, uniformTs>...>;
-
-template <uint32_t set, typename... uniformTs> static consteval
-bool DescriptorSetExistsInUniforms(){ return !FilteredDescriptorPack<set, uniformTs...>::empty; }
-
-template <uint32_t set, typename... uniformTs> static consteval
-uint32_t DescriptorSetCountImpl(){
-	if constexpr (DescriptorSetExistsInUniforms<set, uniformTs...>())
-		return DescriptorSetCountImpl<set + 1, uniformTs...>();
-	else
-		return set;
-}
-
-template <typename... uniformTs> static constexpr
-uint32_t DescriptorSetCount(){ return DescriptorSetCountImpl<0, uniformTs...>(); }
-
-template <uint32_t set, typename... uniformTs> static consteval
-bool AllSmallerSetsExistInUniforms(){
-	if constexpr (set == 0)
-		return DescriptorSetExistsInUniforms<0, uniformTs...>();
-	else
-		return DescriptorSetExistsInUniforms<set, uniformTs..>() && AllSmallerSetsExistInUniforms<set - 1, uniformTs..>();
-}
-
-
-// Attributes
-// -----
-template <VkVertexInputBindingDescription... bindingDescriptions> struct BindingDescriptionPack {};
-template <VkVertexInputAttributeDescription... attributeDescriptions> struct AttributeDescriptionPack {};
-template <typename attributeTP, typename bindingDescriptionP, typename attributeDescriptionP> struct Attributes {};
-template <typename... attributeTs, VkVertexInputBindingDescription... bindingDescriptions, VkVertexInputAttributeDescription... attributeDescriptions>
-struct Attributes<TypePack<attributeTs...>, BindingDescriptionPack<bindingDescriptions...>, AttributeDescriptionPack<attributeDescriptions...>> {
-	
-	static consteval uint32_t TotalSize(){ return (sizeof(attributeTs) + ...); }
-	
-	static_assert((bindingDescriptions.stride + ...) == TotalSize(), "Vertex input binding description strides should sum to the total size of all the attributes.");
-	
-	static_assert(Unique<bindingDescriptions.binding...>(), "Vertex input binding description bindings should be unique.");
-	
-	static_assert(Unique<attributeDescriptions.location...>(), "Vertex input attribute description locations should be unique.");
-	
-	template <typename T> void BindVertexBuffer(){
-		// check binding
-		// ...
-	}
-};
-
-
-// Descriptor set
-// -----
-template <typename descriptorTP> struct DescriptorSet {};
-template <typename... descriptorTs>
-struct DescriptorSet<TypePack<descriptorTs...>> {
-	static constexpr uint32_t descriptorCount = CountT<descriptorTs...>();
-	
-	template <uint32_t index> using iDescriptor = IndexT<index, descriptorTs...>;
-	
-	DescriptorSet(Pipeline &_pipeline, int _index, const DescriptorSetBlueprint &blueprint);
-	~DescriptorSet(){}
-	
-	// For initialisation
-	void InitLayouts();
-	void Update();
-	
-	size_t GetDescriptorCount() const { return descriptors.size(); }
-	std::shared_ptr<Descriptor> GetDescriptor(int index) const { return descriptors[index]; }
-	bool GetUBODynamic() const { return uboDynamicAlignment.has_value(); }
-	const std::optional<VkDeviceSize> &GetUBODynamicAlignment() const { return uboDynamicAlignment; }
-	
-private:
-	Pipeline &pipeline;
-	int index;
-	
-	std::tuple<descriptorTs...> descriptors;
-	
-	// ubo info
-	std::optional<VkDeviceSize> uboDynamicAlignment;
-	
-	// this still relevent? \/\/\/ -
-	/*
-	 `index` is the index of this descriptor set's layout in 'Vulkan::RenderPipeline::descriptorSetLayouts',
-	 
-	 `pipeline.descriptorSetNumber*flightIndex`,
-	 `pipeline.descriptorSetNumber*flightIndex + 1`
-	 ...
-	 `pipeline.descriptorSetNumber*flightIndex + pipeline.descriptorSetNumber - 1`
-	 are the indices of this descriptors flying sets in 'pipeline.descriptorSetsFlying'
-	 */
-	// /\/\/\ -
-};
-
-
-// Uniforms
-// -----
-template <typename uniformTP, typename indexSequenceT> struct UniformsImpl {};
-template <typename... uniformTs, template <uint32_t...> std::integer_sequence indexSequence, uint32_t... indices> struct UniformsImpl<TypePack<uniformTs...>, indexSequence<uint32_t, indices...>> {
-	
-	static_assert(UniformUnique<uniformTs...>(), "No two uniforms should have both matching set and binding.");
-	
-	static_assert((AllSmallerSetsExistInUniforms<uniformTs::setValue, uniformTs...>() && ...), "The union of descriptor set indices for a shader should be consecutive and contain 0.");
-	
-	template <uint32_t set> requires (DescriptorSetExistsInUniforms<set, uniformTs...>())
-	using iDescriptorSet = DescriptorSet<FilteredDescriptorPack<set, uniformTs...>>;
-	
-	static constexpr uint32_t descriptorSetCount = DescriptorSetCount<uniformTs...>();
-	
-	template <uint32_t index>
-	iDescriptorSet<index> &DescriptorSet(){ return std::get<index>(descriptorSets); }
-	
-private:
-	std::tuple<iDescriptorSet<indices>...> descriptorSets;
-};
-
-template <typename... uniformTs> using Uniforms = UniformsImpl<TypePack<uniformTs...>, std::make_integer_sequence<uint32_t, DescriptorSetCount<uniformTs...>()>{}>;
-
-
-// Shader
-// -----
-template <typename PushConstantsT, typename uniformTs...> struct Shader {
-	using pushConstantsPackType = std::conditional_t<std::same_as<PushConstantsT, NoPushConstants>, TypePack<>, TypePack<PushConstantsT>>;
-	using uniformsPackType = TypePack<uniformTs...>;
-};
-template <typename PushConstantsT, typename UniformsT, typename AttributesT>
-struct VertexShader : public Shader<PushConstantsT, UniformsT> {
-	AttributesT attributes;
-};
-
-
-// Shader program
-// -----
-template <typename pushConstantTP, typename UniformsT>
-struct ShaderProgramBase {};
-
-template <typename... pushConstantTs, typename UniformsT>
-struct ShaderProgramBase<TypePack<pushConstantTs...>, UniformsT> {
-	ShaderProgramBase(Interface &_vulkan, const PipelineBlueprint &blueprint);
-	
-	~ShaderProgramBase(){
-		vkDestroyDescriptorPool(vulkan.devices.logicalDevice, descriptorPool, nullptr);
-		vkDestroyPipeline(vulkan.devices.logicalDevice, pipeline, nullptr);
-		vkDestroyPipelineLayout(vulkan.devices.logicalDevice, layout, nullptr);
-		for(VkDescriptorSetLayout &dsl : descriptorSetLayouts){
-			vkDestroyDescriptorSetLayout(vulkan.devices.logicalDevice, dsl, nullptr);
-		}
-	}
-	
-	// ----- Methods to call after Init() -----
-	// Bind the pipeline for subsequent render calls
-	virtual void Bind() = 0;
-	// Set which descriptor sets are bound for subsequent render calls
-	virtual void BindDescriptorSets(int first=0, int number=0, const std::vector<int> &dynamicOffsetNumbers=std::vector<int>()) = 0;
-	void UpdateDescriptorSets(uint32_t first=0); // have to do this every time any elements of any descriptors are changed, e.g. when an image view is re-created upon window resize
-	// Set push constant data
-	template <uint32_t index>
-	void CmdPushConstants(IndexT<index, pushConstantTs...>::type *data){
-		using pushConstantT = IndexT<index, pushConstantTs...>;
-		vkCmdPushConstants(vulkan.commandBuffersFlying[vulkan.currentFrame],
-						   layout,
-						   pushConstantT::stageFlagsValue,
-						   pushConstantT::offsetValue,
-						   sizeof(pushConstantT::type),
-						   data);
-	}
-	
-	// Get the handle of a descriptor set
-	template <uint32_t index>
-	UniformsT::iDescriptorSet<index> &DS(){ return uniforms.DescriptorSet<index>(); }
-	
-protected:
-	Interface &vulkan;
-	
-	UniformsT uniforms;
-	
-	std::vector<VkDescriptorSetLayout> descriptorSetLayouts;
-	std::vector<VkDescriptorSet> descriptorSetsFlying;
-	VkDescriptorPool descriptorPool;
-	
-	std::tuple<pushConstantTs...> pushConstants;
-	
-	VkPipelineLayout layout;
-	VkPipeline pipeline;
-};
-
-template <typename vertexShaderT, typename fragmentShaderT>
-struct GraphicsShaderProgram : public ShaderProgramBase<
-	ConcatenatedPack<vertexShaderT::pushConstantsPackType, fragmentShaderT::pushConstantsPackType>,
-	Uniforms<ConcatenatedPack<vertexShaderT::uniformsPackType, fragmentShaderT::uniformsPackType>>
-> {
-	GraphicsPipeline(Interface &_vulkan, const GraphicsPipelineBlueprint &blueprint);
-	~GraphicsPipeline(){
-		vkDestroyShaderModule(vulkan.devices.logicalDevice, fragShaderModule, nullptr);
-		vkDestroyShaderModule(vulkan.devices.logicalDevice, vertShaderModule, nullptr);
-	}
-	
-	void Bind() override;
-	void BindDescriptorSets(int first=0, int number=0, const std::vector<int> &dynamicOffsetNumbers=std::vector<int>()) override;
-	
-private:
-	vertexShaderT vertexShader;
-	fragmentShaderT fragmentShader;
-	
-//			VkShaderModule vertShaderModule; ?
-//			VkShaderModule fragShaderModule; ?
-};
-
-template <typename computeShaderT>
-struct ComputeShaderProgram : public ShaderProgramBase<computeShaderT::pushConstantsPackType> {
-	ComputePipeline(Interface &_vulkan, const ComputePipelineBlueprint &blueprint);
-	~ComputePipeline(){
-		vkDestroyShaderModule(vulkan.devices.logicalDevice, shaderModule, nullptr);
-	}
-	
-	void Bind() override;
-	void BindDescriptorSets(int first=0, int number=0, const std::vector<int> &dynamicOffsetNumbers=std::vector<int>()) override;
-	
-private:
-	computeShaderT computeShader;
-	
-//		VkShaderModule shaderModule; ?
-};
 
 
 
@@ -944,12 +309,12 @@ private:
 	uint32_t currentFrame = 0;
 	
 	// ----- structures -----
-	std::vector<std::optional<UniformBufferObject>> uniformBufferObjects;
-	std::vector<std::optional<StorageBufferObject>> storageBufferObjects;
-	std::vector<std::optional<TextureImage>> textureImages;
-	std::vector<VkSampler> textureSamplers;
-	std::vector<std::optional<BufferedRenderPass>> bufferedRenderPasses;
-	std::vector<std::optional<LayeredBufferedRenderPass>> layeredBufferedRenderPasses;
+//	std::vector<std::optional<UniformBufferObject>> uniformBufferObjects;
+//	std::vector<std::optional<StorageBufferObject>> storageBufferObjects;
+//	std::vector<std::optional<TextureImage>> textureImages;
+//	std::vector<VkSampler> textureSamplers;
+//	std::vector<std::optional<BufferedRenderPass>> bufferedRenderPasses;
+//	std::vector<std::optional<LayeredBufferedRenderPass>> layeredBufferedRenderPasses;
 	
 	// vertex / index buffers
 	std::vector<std::optional<VertexBufferObject>> vertexBufferObjects;
