+++ include/Descriptors/StorageImages.hpp
@@ -1,0 +1,22 @@
+#pragma once
+
+#include "DescriptorBase.hpp"
+
+namespace EVK {
+
+template <uint32_t binding, VkShaderStageFlags stageFlags>
+class StorageImagesDescriptor : public DescriptorBase<binding, stageFlags> {
+public:
+	StorageImagesDescriptor(const std::vector<int> &_indices) : indices(_indices) {}
+	
+	VkDescriptorSetLayoutBinding LayoutBinding() const override;
+	
+	VkWriteDescriptorSet DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const override;
+	
+	VkDescriptorPoolSize PoolSize() const override;
+	
+private:
+	std::vector<int> indices;
+};
+
+} // namespace EVK
