+++ include/Descriptors/TextureSamplers.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+#include "DescriptorBase.hpp"
+
+namespace EVK {
+
+template <uint32_t binding, VkShaderStageFlags stageFlags>
+class TextureSamplersDescriptor : public DescriptorBase<binding, stageFlags> {
+public:
+	TextureSamplersDescriptor(const std::vector<int> &_indices) : indices(_indices) {}
+	
+	VkDescriptorSetLayoutBinding LayoutBinding() const override;
+	
+	VkWriteDescriptorSet DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const override;
+	
+	VkDescriptorPoolSize PoolSize() const override;
+	
+private:
+	std::vector<int> indices;
+};
+
+} // namespace EVK
+
