+++ include/Static.hpp
@@ -9,6 +9,8 @@
 #include <tuple>
 #include <type_traits>
 
+#include "Header.hpp"
+
 namespace Static {
